src/actuators/drivers.py
#!/usr/bin/env python3

import rospy
from std_msgs.msg import Float64
from sensor_msgs.msg import JointState

class ActuatorDriver:
    """Base class for actuator drivers"""
    def __init__(self, name, joint_name):
        self.name = name
        self.joint_name = joint_name
        self.position = 0.0
        self.velocity = 0.0
        self.effort = 0.0

    def set_position(self, position):
        """Set actuator position"""
        self.position = position

    def get_position(self):
        """Get current position"""
        return self.position

class ServoDriver(ActuatorDriver):
    """Driver for servo motors"""
    def __init__(self, name, joint_name, pin):
        super().__init__(name, joint_name)
        self.pin = pin
        self.min_pulse = 500
        self.max_pulse = 2500
        
    def map_angle_to_pulse(self, angle):
        """Map angle (-90 to 90) to pulse width"""
        return int(((angle + 90) / 180) * (self.max_pulse - self.min_pulse) + self.min_pulse)

class DCMotorDriver(ActuatorDriver):
    """Driver for DC motors"""
    def __init__(self, name, joint_name, pwm_pin, dir_pin):
        super().__init__(name, joint_name)
        self.pwm_pin = pwm_pin
        self.dir_pin = dir_pin
        self.speed = 0
        
    def set_speed(self, speed):
        """Set motor speed (-1 to 1)"""
        self.speed = max(min(speed, 1.0), -1.0)

class StepperDriver(ActuatorDriver):
    """Driver for stepper motors"""
    def __init__(self, name, joint_name, step_pin, dir_pin, enable_pin):
        super().__init__(name, joint_name)
        self.step_pin = step_pin
        self.dir_pin = dir_pin
        self.enable_pin = enable_pin
        self.steps_per_rev = 200
        
    def move_steps(self, steps):
        """Move specified number of steps"""
        pass  # Implement actual stepping logic 