requirements.txt
rosdep>=0.22.1
rospkg>=1.5.0
catkin-pkg>=0.5.2
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
PyYAML>=5.4.1
transforms3d>=0.3.1
pyserial>=3.5
opencv-python>=4.5.0
RPi.GPIO>=0.7.0
smbus2>=0.4.1
tf2_ros>=0.7.5
sensor_msgs>=1.13.1
geometry_msgs>=1.13.1
visualization_msgs>=1.13.1 